frequent_traders_detection/requirements.txt
confluent-kafka==2.3.0
python-decouple==3.8
redis==5.0.1
websockets==12.0
